Remove and centralize device functions
