+++ server/test/utils/device.test.js
@@ -3,6 +3,7 @@ const {
   getDeviceParam,
   setDeviceParam,
   getDeviceFeature,
+  setDeviceFeature,
   hasDeviceChanged,
   mergeFeatures,
   mergeDevices,
@@ -132,6 +133,54 @@ describe('getDeviceFeature', () => {
   it('should return null on empty device', () => {
     const value = getDeviceFeature({}, 'unknown-category', 'unknown-type');
     expect(value).to.equal(null);
+  });
+});
+
+describe('setDeviceFeature', () => {
+  it('should create array and add feature', () => {
+    const device = {};
+    const feature = {
+      selector: 'selector',
+    };
+
+    setDeviceFeature(device, feature);
+
+    expect(device).deep.eq({
+      features: [feature],
+    });
+  });
+
+  it('should add feature to array', () => {
+    const device = { features: [{ selector: 'no-match' }] };
+    const feature = {
+      selector: 'selector',
+    };
+
+    setDeviceFeature(device, feature);
+
+    expect(device).deep.eq({
+      features: [{ selector: 'no-match' }, feature],
+    });
+  });
+
+  it('should replace feature into array', () => {
+    const device = {
+      features: [
+        {
+          selector: 'selector',
+        },
+      ],
+    };
+    const feature = {
+      name: 'name',
+      selector: 'selector',
+    };
+
+    setDeviceFeature(device, feature);
+
+    expect(device).deep.eq({
+      features: [feature],
+    });
   });
 });
