+++ server/utils/device.js
@@ -69,6 +69,31 @@ function getDeviceFeature(device, category, type) {
   return null;
 }
 
+/**
+ * @description Add or update a param to a device.
+ * @param {object} device - Device to add parameter.
+ * @param {object} feature - The feature to add.
+ * @returns {object} The device.
+ * @example
+ * setDeviceFeature({ features: [] }, { selector: 'feature' })
+ */
+function setDeviceFeature(device, feature) {
+  let { features } = device;
+  if (!features) {
+    features = [];
+    device.features = features;
+  }
+
+  const featureIndex = features.findIndex((p) => p.selector === feature.selector);
+  if (featureIndex >= 0) {
+    features[featureIndex] = feature;
+  } else {
+    features.push(feature);
+  }
+
+  return device;
+}
+
 const matchFeature = (features, feature) => {
   return features.findIndex((f) => f.external_id === feature.external_id);
 };
@@ -204,6 +229,7 @@ module.exports = {
   getDeviceParam,
   setDeviceParam,
   getDeviceFeature,
+  setDeviceFeature,
   hasDeviceChanged,
   mergeFeatures,
   mergeDevices,
